modifying qualification for users
