+++ care/emr/resources/encounter/spec.py
@@ -2,11 +2,12 @@
 import datetime
 
 from django.utils import timezone
-from pydantic import UUID4, BaseModel, model_validator
+from pydantic import UUID4, BaseModel
 
 from care.emr.models import Encounter, EncounterOrganization, TokenBooking
 from care.emr.models.patient import Patient
 from care.emr.resources.base import EMRResource
+from care.emr.resources.common.period import PeriodSpec
 from care.emr.resources.encounter.constants import (
     AdmitSourcesChoices,
     ClassChoices,
@@ -21,17 +22,6 @@ from care.emr.resources.patient.spec import PatientListSpec
 from care.emr.resources.scheduling.slot.spec import TokenBookingReadSpec
 from care.emr.resources.user.spec import UserSpec
 from care.facility.models import Facility
-
-
-class PeriodSpec(BaseModel):
-    start: datetime.datetime | None = None
-    end: datetime.datetime | None = None
-
-    @model_validator(mode="after")
-    def validate_period(self):
-        if (self.start and self.end) and (self.start > self.end):
-            raise ValueError("Start Date cannot be greater than End Date")
-        return self
 
 
 class HospitalizationSpec(BaseModel):
